requirements.txt
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
pytest
unittest